test: make the test independent of others
This test fast failing with 'source not well formed' error message,
which is Weaviate's speak for 'I dont think this property is a reference'.
Still not sure why, but I figured it may have to do with how the test-suite is structured;
namely, that each test relies on the state produced/left by the previous test.

To test reference deletion, we instead simply create 2 objects,
reference one from the other and then delete that reference.
No external state.